api/requirements.txt
azure-functions==1.21.3
azure-cosmos==4.7.0
azure-identity==1.19.0
openai==1.59.6
